+++ e2e/commands/import.e2e.js
@@ -800,6 +800,22 @@ describe('bit import', function () {
         expect(result.trim()).to.equal('got is-type and got is-string and got foo');
       });
     });
+    describe('change the dependency version in bit.json', () => {
+      before(() => {
+        helper.getClonedLocalScope(clonedLocalScope);
+        helper.importComponent('bar/foo --conf');
+        const bitJsonPath = path.join(helper.localScopePath, 'components/bar/foo/bit.json');
+        const bitJson = helper.readBitJson(bitJsonPath);
+        bitJson.dependencies[`${helper.remoteScope}/utils/is-string`] = '0.0.2';
+        helper.writeBitJson(bitJson, bitJsonPath);
+      });
+      it('bit show should display the dependency version according to bit.json', () => {
+        const output = helper.showComponentParsed('bar/foo -c');
+        expect(output.componentFromFileSystem.dependencies[0].id).to.equal(
+          `${helper.remoteScope}/utils/is-string@0.0.2`
+        );
+      });
+    });
   });
 
   describe('components with auto-resolve dependencies using css', () => {